requirements.txt
setuptools
setuptools_git

requests

Flask
pyyaml

wtforms
Flask_WTF

psycopg2

